requirements.txt
streamlit==1.28.0
easyocr==1.7.0
pytesseract==0.3.10
Pillow==9.5.0
numpy==1.24.3
opencv-python-headless==4.8.0.74
